+++ src/ebisp/std.h
@@ -1,0 +1,6 @@
+#ifndef STD_H_
+#define STD_H_
+
+void load_std_library(Gc *gc, struct Scope *scope);
+
+#endif  // STD_H_
